requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
anthropic==0.3.6
anyio==3.7.1
appnope==0.1.3
asgiref==3.5.0
astroid==2.15.5
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
boltons==23.0.0
camel-converter==3.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chevron==0.14.0
click==8.1.6
continuedev @ file:///Users/jakeascher/.continue/server/continuedev-0.1.2-py3-none-any.whl#sha256=3a617955111839fe0313c48835a456de5d864fd00e25323e30b33e49af2d0ed3
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.13
decorator==5.1.1
Deprecated==1.2.14
diff-match-patch==20230430
dill==0.3.6
directory-tree==0.0.3.1
distro==1.8.0
Django==4.0.4
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.95.2
fonttools==4.40.0
frozenlist==1.4.0
fsspec==2023.6.0
gpt-index==0.6.38.post1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
jsonref==1.1.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.245
langsmith==0.0.14
lazy-object-proxy==1.9.0
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
meilisearch-python-async==1.4.8
monotonic==1.6
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
posthog==3.0.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
PyGithub==1.59.0
Pygments==2.15.1
PyJWT==2.8.0
pylint==2.17.4
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rpds-py==0.9.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.2
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
